\name{ordPens-internal}
\title{Internal ordPens functions}
\alias{cd}
\alias{coding}
\alias{genRidge}
\alias{ordAOV1}
\alias{ordAOV2}
\description{Internal ordPens functions}
\usage{
cd(x)
coding(x, constant=TRUE, splitcod=TRUE)
genRidge(x, y, offset, omega, lambda, model, delta=1e-6, maxit=25)
ordAOV1(x, y, type, nsim, ...)
ordAOV2(x, y, type, nsim, ...)
}
\author{Jan Gertheiss}
\details{
  These are not to be called by the user.
}
\keyword{internal}
