\name{ordPens-internal}
\title{Internal ordPens functions}
\alias{coding}
\alias{genRidge}
\description{Internal ordPens functions}
\usage{
coding(x, constant=TRUE, splitcod=TRUE)
genRidge(x, y, offset, omega, lambda, model, delta=1e-6, maxit=25) 
}
\author{Jan Gertheiss}
\details{
  These are not to be called by the user.
}
\keyword{internal}
